+++ client-app/src/app/available-assets-list/available-assets-list.ts
@@ -80,6 +80,9 @@ export class AvailableAssetsList {
   }
  
   deleteAsset(id: any) {
+    if (!confirm('Are you sure you want to delete this Asset?')) {
+      return;
+    }
     this.assetService.deleteAsset(id).subscribe(() => {
       this.loadAssets(); 
     });
